# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Zhalwadi Samaj API Documentation",
        "description": "API documentation for Zhalwadi Samaj mobile application",
        "contact": {
            "email": "support@zhalwadisamaj.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://parivar.test",
            "description": "API Server"
        }
    ],
    "paths": {
...
        "/api/member/search_member": {
            "get": {
                "tags": [
                    "General"
                ],
                "summary": "Search Members",
                "description": "Search members with various filters",
                "operationId": "6a279665104c06b81ac41f5d0b93be31",
                "parameters": [
                    {
                        "name": "search_key",
                        "in": "query",
                        "description": "Search key (name, unique_number, native, profession)",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "search_value",
                        "in": "query",
                        "description": "Search value",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "filter_by_blood_group",
                        "in": "query",
                        "description": "Filter by blood group",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "filter_by_status",
                        "in": "query",
                        "description": "Filter by status",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "length",
                        "in": "query",
                        "description": "Number of results per page",
                        "schema": {
                            "type": "integer",
                            "default": 20
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Members found successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/job/get_all": {
            "get": {
                "tags": [
                    "Jobs"
                ],
                "summary": "Get All Jobs",
                "operationId": "223c447a111876c5a01534568e03d600",
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer",
                "bearerFormat": "JWT",
                "description": "Enter your bearer token in the format: Bearer {token}"
            }
        }
    },
...
    "security": [
        {
            "bearerAuth": []
        }
    ]
}